Create ossar.yml
This is an additional security check to make sure our code is secure. This check is not certified by GitHub and are instead provided by a third-party. This check may not be necessary and we can turn this off later if we want to, but for now I am adding it in.