Adding Starting Files to Begin Testing the Repo
I was trying to merge the JacobW branch to the main branch with my added GitHub security checks, but because I didn't have any code files for these checks to, well check, all of the tests failed and the repo would not let me commit the changes. 

So, now I've added some files for the checks to now test against, although only one of the four files is a .cpp file. The other files are the two "books" files that we'll need to use in our project and the Documentation Paperwork for this project.