Adding an Actual VS Project Solution to the Repo
Well, after adding a .cpp file I still couldn't get the security checks to pass. This time it is because it says that it was not able to build my code since it was just the one code file and no configurations or anything of the sort.

So, now I will be trying to run the checks again with an actual VS Project and Solution in the directory. Also, just looking through how Git is "seeing" the current changes, it seems to only be 4 files getting added to the repo, with none of the files from the .vs folder showing up here at all! So that's at least promising for the current .gitignore file!