Update codeql.yml
Ok, well adding those two extra OSs to the runs-on array made the check stall b/c it was trying to find a "runner" that would run the checks for all 3 of the stated OSs at once. Instead now I am removing the default "ubuntu-latest" and the "macos-latest" from the array so that it will only try to run on windows. Let's see if this will make it work finally.