Update ossar.yml
For this open source third party security check, there does not seem to be much for me to be able to change to try to get this check to work. If this one change that I make doesn't work, then I am just going to delete this from the repo and only use the GitHub security check.