Completed Implementing the Book Class
For now I have made all of the properties strings, even the ISPN and the publication year. Mainly doing this since the source for these properties are going to be coming in as strings and our output onto the screen is going to be strings. That being said, we may still change these two properties to unsigned ints later.

Also all of the setter functions for the properties are private, so once they are set by the constructor they cannot be changed. If we ever need to change them in the future though, we can change these functions to be public.