Fixing Back-end's sign in function
Before, the function would only test the very first option in the users.csv file and then return true or false depending on if the input matched it or not. 

This change makes it so that the for loop will go through and check the entire list first before returning false, so basically doing what it was meant to do in the first place. And when it does find a match it returns and breaks out of the loop early.