Adding a Menu System to the Front End
Basically getting rid of all of the "cin >> input;" lines and replacing them with getline(cin, input);

Also since our program in now expected to do multiple tasks based on the user's request, I have overhauled the Front End to make use of basically a Menu System. It is currently still a work-in-progress, but I want to upload it for us to at least have a start with it.