Completing the work that was due yesterday
Made the getter functions in the Book class const functions in order to get them to work properly with multisets (what I used for the user's shopping list)

Made the shopping list functionality for this week and the menu for it.

Added a compareBooksByMSRP() function in the utilities file to be able to compare books by prices.