Updating Database to be more stable
So I found an issue in my database design that allowed for a few columns to be NULL when we always expect them to have a valid value. This lead to a few bad records sneaking into the database and causing problems. 

This commit is me deleting those records and remaking the BOOKS table to not allow NULLs in any of our columns except for Description and Genre.

Also updating the Gantt chart on my end to reflect the work that Mike did today.