Various Bugfixes
Fixed so that if user enters nothing for making a new shopper it doesn't crash.

Removed placeholder console message artifact.

Fixed it so that the user's shopping list gets cleared after it has been purchased.

Found that the database is tracking the shoppers' total spent to a percent of a cent rather than just a cent. This was not fixed as was instead recorded as an assumption.