Integrating Most of our Previous Code
So far I have included all of the files that I believe will work with QT right away. The final files I do not know if they will work so I am leaving them out for the next commit. These files are the backEnd.h, backEnd.cpp, and the Front End cpp file. The first two are like the rest where they will be mostly copying and pasting. the final file that handles our front end is a different story, I am just going to make a block comment in our new front end on all of the things we will need to add to our front end based on what is in our old front end.